requirements.txt
aws-cdk.core==1.72.0
aws-cdk.aws_s3==1.72.0
aws-cdk.aws-lambda==1.72.0
aws_cdk.aws_lambda_event_sources==1.72.0
